Refactor derived `OperationData` classes to a separate header
